requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
DAWG-Python==0.7.2
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.6
django-celery-results==2.5.1
djangorestframework==3.15.2
docopt==0.6.2
drf-yasg==1.21.8
inflection==0.5.1
kombu==5.4.2
numpy==2.2.3
packaging==24.2
prompt_toolkit==3.0.50
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
regex==2024.11.6
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13